Split FileLogger to its own file
