+++ packages/test/test-service-load/src/utils.ts
@@ -3,12 +3,10 @@
  * Licensed under the MIT License.
  */
 
-import crypto from "crypto";
 import fs from "fs";
 
 import {
 	DriverEndpoint,
-	ITelemetryBufferedLogger,
 	ITestDriver,
 	TestDriverTypes,
 } from "@fluid-internal/test-driver-definitions";
@@ -22,17 +20,12 @@ import { IContainer, IFluidCodeDetails } from "@fluidframework/container-definit
 // eslint-disable-next-line import/no-deprecated
 import { type IDetachedBlobStorage, Loader } from "@fluidframework/container-loader/internal";
 import { IContainerRuntimeOptions } from "@fluidframework/container-runtime/internal";
-import {
-	ConfigTypes,
-	IConfigProviderBase,
-	ITelemetryBaseEvent,
-	LogLevel,
-} from "@fluidframework/core-interfaces";
-import { assert, LazyPromise } from "@fluidframework/core-utils/internal";
+import { ConfigTypes, IConfigProviderBase } from "@fluidframework/core-interfaces";
+import { assert } from "@fluidframework/core-utils/internal";
 import { ICreateBlobResponse } from "@fluidframework/driver-definitions/internal";
-import { createChildLogger } from "@fluidframework/telemetry-utils/internal";
 import { LocalCodeLoader } from "@fluidframework/test-utils/internal";
 
+import { FileLogger } from "./FileLogger.js";
 import { ILoadTest, createFluidExport, type IRunConfig } from "./loadTestDataStore.js";
 import {
 	generateConfigurations,
@@ -44,94 +37,6 @@ import { pkgName, pkgVersion } from "./packageVersion.js";
 import { ILoadTestConfig, ITestConfig } from "./testConfigFile.js";
 
 const packageName = `${pkgName}@${pkgVersion}`;
-
-class FileLogger implements ITelemetryBufferedLogger {
-	private static readonly loggerP = new LazyPromise<FileLogger>(async () => {
-		if (process.env.FLUID_TEST_LOGGER_PKG_SPECIFIER !== undefined) {
-			await import(process.env.FLUID_TEST_LOGGER_PKG_SPECIFIER);
-			const logger = getTestLogger?.();
-			assert(logger !== undefined, "Expected getTestLogger to return something");
-			return new FileLogger(logger);
-		} else {
-			return new FileLogger(undefined);
-		}
-	});
-
-	public static async createLogger(dimensions: {
-		driverType: string;
-		driverEndpointName: string | undefined;
-		profile: string;
-		runId: number | undefined;
-	}) {
-		return createChildLogger({
-			logger: await this.loggerP,
-			properties: {
-				all: dimensions,
-			},
-		});
-	}
-
-	public static async flushLogger(runInfo?: { url: string; runId?: number }) {
-		await (await this.loggerP).flush(runInfo);
-	}
-
-	private error: boolean = false;
-	private readonly schema = new Map<string, number>();
-	private logs: ITelemetryBaseEvent[] = [];
-	public readonly minLogLevel: LogLevel = LogLevel.verbose;
-
-	private constructor(private readonly baseLogger?: ITelemetryBufferedLogger) {}
-
-	async flush(runInfo?: { url: string; runId?: number }): Promise<void> {
-		const baseFlushP = this.baseLogger?.flush();
-
-		if (this.error && runInfo !== undefined) {
-			const logs = this.logs;
-			const outputDir = `${__dirname}/output/${crypto
-				.createHash("md5")
-				.update(runInfo.url)
-				.digest("hex")}`;
-			if (!fs.existsSync(outputDir)) {
-				fs.mkdirSync(outputDir, { recursive: true });
-			}
-			// sort from most common column to least common
-			const schema = [...this.schema].sort((a, b) => b[1] - a[1]).map((v) => v[0]);
-			const data = logs.reduce(
-				(file, event) =>
-					// eslint-disable-next-line @typescript-eslint/no-base-to-string
-					`${file}\n${schema.reduce((line, k) => `${line}${event[k] ?? ""},`, "")}`,
-				schema.join(","),
-			);
-			const filePath = `${outputDir}/${runInfo.runId ?? "orchestrator"}_${Date.now()}.csv`;
-			fs.writeFileSync(filePath, data);
-		}
-		this.schema.clear();
-		this.error = false;
-		this.logs = [];
-		return baseFlushP;
-	}
-	send(event: ITelemetryBaseEvent): void {
-		if (typeof event.testCategoryOverride === "string") {
-			event.category = event.testCategoryOverride;
-		} else if (
-			typeof event.message === "string" &&
-			event.message.includes("FaultInjectionNack")
-		) {
-			event.category = "generic";
-		}
-		this.baseLogger?.send({ ...event, hostName: pkgName, testVersion: pkgVersion });
-
-		event.Event_Time = Date.now();
-		// keep track of the frequency of every log event, as we'll sort by most common on write
-		Object.keys(event).forEach((k) => this.schema.set(k, (this.schema.get(k) ?? 0) + 1));
-		if (event.category === "error") {
-			this.error = true;
-		}
-		this.logs.push(event);
-	}
-}
-
-export const createLogger = FileLogger.createLogger.bind(FileLogger);
 
 const codeDetails: IFluidCodeDetails = {
 	package: packageName,
@@ -189,7 +94,7 @@ export async function initialize(
 		generateConfigurations(seed, optionsOverride?.configurations),
 	);
 
-	const logger = await createLogger({
+	const logger = await FileLogger.createLogger({
 		driverType: testDriver.type,
 		driverEndpointName: testDriver.endpointName,
 		profile: profileName,
